refactor: move elfdynlib and machodylib under dynlib/ hierarchy
Move internal/elfdynlib/ and internal/machodylib/ into
internal/dynlib/elfdynlib/ and internal/dynlib/machodylib/ to mirror
the existing security/{elfanalyzer,machoanalyzer} structure.

dynlib/ now serves as the common parent: shared error types at the top
level, format-specific packages as sub-packages.
