+++ internal/dynlib/elfdynlib/analyzer.go
