+++ internal/dynlib/elfdynlib/analyzer_test.go
