+++ internal/dynlib/elfdynlib/default_paths.go
