+++ internal/dynlib/elfdynlib/errors.go
