+++ internal/dynlib/elfdynlib/ldcache.go
