+++ internal/dynlib/elfdynlib/resolver.go
