+++ internal/dynlib/elfdynlib/verifier.go
