+++ internal/dynlib/machodylib/analyzer.go
