+++ internal/dynlib/machodylib/dyld_extractor.go
