+++ internal/dynlib/machodylib/errors.go
