+++ internal/dynlib/machodylib/testdata/dyld_headers/mach-o/fixup-chains.h
@@ -331,4 +331,3 @@ struct dyld_chained_import_addend64
 };
 
 #endif // __MACH_O_FIXUP_CHAINS__
-
